Revert "refactor(engine): use hw event callback in a thread-safe way (#9335)"
This reverts commit ce3550af513cfab6b654d27061f674d3da7e468f.
