+++ api/tests/opentrons/protocol_runner/test_thread_async_queue.py
@@ -9,7 +9,7 @@ from typing import List, NamedTuple
 
 import pytest
 
-from opentrons.thread_async_queue import (
+from opentrons.protocol_runner.thread_async_queue import (
     ThreadAsyncQueue,
     QueueClosed,
 )
